Don't let Net::SNMP translate strings to hex
Net::SNMP's default behavior is to return printable
strings as-is, and translate non-printable strings to
hex.  Once we disabled the lower data rates, this
list of rates became all printable, so we stopped
being able to parse it.  Turn off the translation,
so that we always just get a list of characters,
and use ord() instead of hex() to translate them.